Issue #4587: Moved unused resource classes
