--- src/test/resources/com/puppycrawl/tools/checkstyle/checks/InputAstRegressionNewlineCrAtEndOfFile.java
@@ -1,1 +1,0 @@
-////////////////////////////////////////////////////////////////////////////////// Test case file for checkstyle.// Created: 2015////////////////////////////////////////////////////////////////////////////////package com.puppycrawl.tools.checkstyle.checks;/** * Test case for detection of an existing CR newline at EOF, using the  * NewlineAtEndOfFileCheck. * @author Martin Steiger **/public interface InputAstRegressionNewlineCrAtEndOfFile{/** Test with ending with CR newline*/}
