+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/imports/unusedimports/Input15Extensions.java
@@ -1,5 +1,5 @@
 // someexamples of 1.5 extensions
-package com.puppycrawl.tools.checkstyle.checks;
+package com.puppycrawl.tools.checkstyle.checks.imports.unusedimports;
 
 @interface MyAnnotation1 {
     String name();
