+++ src/test/resources/com/puppycrawl/tools/checkstyle/grammars/InputAstRegressionNewlineCrAtEndOfFile.java
@@ -1,0 +1,1 @@
+////////////////////////////////////////////////////////////////////////////////// Test case file for checkstyle.// Created: 2015////////////////////////////////////////////////////////////////////////////////package com.puppycrawl.tools.checkstyle.grammars;/** * Test case for detection of an existing CR newline at EOF, using the  * NewlineAtEndOfFileCheck. * @author Martin Steiger **/public interface InputAstRegressionNewlineCrAtEndOfFile{/** Test with ending with CR newline*/}
