+++ src/test/resources/com/puppycrawl/tools/checkstyle/grammars/InputNewlineCrAtEndOfFileAst.txt
@@ -16,8 +16,8 @@ PACKAGE_DEF -> package [5:0]
 |   |   |   |   `--IDENT -> puppycrawl [5:12]
 |   |   |   `--IDENT -> tools [5:23]
 |   |   `--IDENT -> checkstyle [5:29]
-|   `--IDENT -> checks [5:40]
-`--SEMI -> ; [5:46]
+|   `--IDENT -> grammars [5:40]
+`--SEMI -> ; [5:48]
 INTERFACE_DEF -> INTERFACE_DEF [12:0]
 |--MODIFIERS -> MODIFIERS [12:0]
 |   |--BLOCK_COMMENT_BEGIN -> /* [7:0]
